refactor: Move theme related code to utils subfolder.
