+++ app/routes/action+/set-theme.ts
@@ -1,3 +1,3 @@
-import { createThemeAction } from '#.server/theme';
+import { createThemeAction } from '#utils/theme/theme.server';
 
 export const action = createThemeAction();
