+++ app/utils/theme/client-hints.server.ts
