+++ app/utils/theme/media-query-fallback.tsx
@@ -1,4 +1,4 @@
-import { useTheme } from './theme';
+import { useTheme } from './theme.provider';
 
 const clientThemeCode = `
 (() => {
